refactor(wallet)!: move cachedGetPassword under KeyManagement.util
