+++ packages/wallet/src/KeyManagement/index.ts
@@ -5,4 +5,3 @@ export * from './restoreKeyAgent';
 export * as util from './util';
 export * from './emip3';
 export * from './types';
-export * from './cachedGetPassword';
