+++ packages/wallet/src/KeyManagement/util/index.ts
@@ -1,3 +1,4 @@
 export * from './bip39';
 export * from './key';
 export * from './ownSignatureKeyPaths';
+export * from './cachedGetPassword';
